Move navigation component into a seperate file, so that it doesn't interfere with the snapshot tests in App
